Update the check for update entries more detailed
By only checking length then feeds good potentially always have a length
of 5 but content being different, because there is more than 5 posts
every 2 weeks. When now check the title of the latest post to see if its
the same as the one in the cache, if not then display the notification
stating there is updated content.